fix: Update tables ids as bigint

+++ src/main/resources/db/migration/V3__create-table-usuario.sql
@@ -1,13 +1,13 @@
 CREATE TABLE tb_usuario (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     nome VARCHAR(255) NOT NULL,
     email VARCHAR(255) NOT NULL UNIQUE,
     senha VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE tb_perfil_usuario (
-    usuario_id INT,
-    perfil_id INT,
+    usuario_id BIGINT,
+    perfil_id BIGINT,
     PRIMARY KEY (usuario_id, perfil_id),
     FOREIGN KEY (usuario_id) REFERENCES tb_usuario(id) ON DELETE CASCADE,
     FOREIGN KEY (perfil_id) REFERENCES tb_perfil(id) ON DELETE CASCADE
